Unify metricdata type transforms into one file
